Move asset_path methods out of Config
These methods are nicely isolated in a small AssetPath module
